GitHub Actionsの設定

+++ db/migrate/20240216040305_remove_video_id_from_bookmarks.rb
@@ -1,5 +1,7 @@
 class RemoveVideoIdFromBookmarks < ActiveRecord::Migration[7.1]
-  def change
+  if column_exists?(:bookmarks, :video_id)
+    def change
     remove_column :bookmarks, :video_id, :string
+    end
   end
 end
